+++ NEWS.md
@@ -1,4 +1,4 @@
-# weird (development version)
+# weird 3.1.0
 
 - `dist_kde()` HDR calculations now always use the density at the data points (previously this required at least 200 observations), following `hdrcde`.
 - `dist_density()` removed in favour of `distributional::dist_density()`.